Fix DTO duplicate MediaFileDto, add missing Repository methods (countByCreatedAtAfter, findByCreatedAtAfter), fix RedisService generic get() method, fix AdminStatsService imports

+++ spring_boot_backend/src/main/kotlin/com/androidinsta/Repository/User/CommentRepository.kt
@@ -62,5 +62,5 @@ interface CommentRepository : JpaRepository<Comment, Long> {
     /**
      * Đếm số replies của một comment
      */
-    fun countByParentCommentId(parentCommentId: Long): Int
+    fun countByParentCommentId(parentCommentId: Long): Long
 }
